Make ConfigElement into trait
This patch rewrites the `ConfigElement` enum to be a trait for accessing
a configuration element object.

The trait implements `downcast_rs::DowncastSync` so we can provide
typed access to the actual objects that were loaded from the interface
of the `Config` type.

This is neccessary to reach as much format independence as possible. We
need to be able to access certain value types so that we can implement
proper layering in the crate. For that, we need access to atoms (bool,
integer types, string) and composite types (list, map).
The `ConfigElement` trait helps us.

The `is_` functions in the trait with the corrosponding `as_` function
from the trait and `.is_some()`.

The List and Object types are abstracted via traits as well.

The rest of the crate was adapted as much as necessary to be able to work
with the new interface.
The `IntoConfigElement` trait is no longer necessary and thus was
removed.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>