Optimize implementation
This patch optimizes the match implementation in the access function. We
introduce a type ConfigElementType, which is used internally to describe
the element type.
The ConfigElement::get_type() function is auto-implemented using the
ConfigElement::is_* functions. Its implementation is easily verifyable
(more easily than the more complex match in ConfigElement::access()).

The big match block in ConfigElement::access() uses
ConfigElement::get_type() and we let the compiler check that all
patterns are covered.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>