Simplify interface using a helper type
This reduces the API surface by introducing the `Layers` helper type.

The actual work is done on the `Layers` type now. I first tried to
abstract this with a (sealed) helper trait, but this feels clunky. So I
figured that calling a `.layers()` function to get access to the
configuration layers is not too much of a burden. The `Config` and the
`AsyncConfig` type provide access to `Layers`, but we do not have to
make weird things for both of these types to essentially provide the
same interfaces for getting access to the actual configuration values.

As said, having these abstract over a trait just felt clunky, so this is
(IMHO) a better interface.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>