Add test for nested custom config types
With this patch we see that a user of the crate needs to implement
FromConfigElement for their custom type as well if they want to nest
types.

I am not yet sure whether this is the way to go, but for the first step,
this is sufficient. I think we want to add a derive macro for this as
well, though.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>